Refactor: DELETE API 공통 컴포넌트화
