+++ src/common/ImageUploadAPI.jsx
@@ -1,5 +1,5 @@
 import { BASE_URL } from './BASE_URL';
-import ImageResize from './ImageResize';
+import imageResizeAPI from './ImageResizeAPI';
 
 // postImage라는 API를 부를때 file을 넘기는데
 // 이때 file이 null일 수 있다.
@@ -11,7 +11,7 @@ async function postImage(file) {
 
     // jpeg, png인 경우에만 압축을 진행
     if (newFile.type === 'image/jpeg' || newFile.type === 'image/png') {
-      newFile = await ImageResize(file);
+      newFile = await imageResizeAPI(file);
     } else if (file.size >= 10 * 1024 * 1024) {
       alert('10MB 이하의 이미지를 올려주세요');
       new Error('용량 초과');
